MainPage switchs Login/Register formulary
