+++ calorieburningapp.client/src/Components/Card/Card.css
@@ -1,5 +1,5 @@
 .card {
-    min-width: 350px;
+    min-width: 300px;
     min-height: 200px;
     background-color: white;
     border-radius: 13px;
