+++ calorieburningapp.client/src/Components/MainPage/Login/index.js
@@ -1,0 +1,1 @@
+export { default } from './Login';
