--- calorieburningapp.client/src/Components/MainPage/RegisterLogin.css
@@ -1,49 +1,0 @@
-.register-login {
-    display: flex;
-    justify-content: space-between;
-  }
-  
-  .main-content {
-    flex: 1;
-    padding: 20px;
-  }
-  
-  .login-form {
-    flex: 1;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-  }
-  
-  .login-form input {
-    margin-bottom: 10px;
-    padding: 8px;
-    width: 250px;
-  }
-  
-  .login-button {
-    background-color: green;
-    color: white;
-    padding: 10px 20px;
-    border: none;
-    border-radius: 5px;
-    cursor: pointer;
-    margin-bottom: 10px;
-  }
-  
-  .register-button {
-    background-color: blue;
-    color: white;
-    padding: 10px 20px;
-    border: none;
-    border-radius: 5px;
-    cursor: pointer;
-    margin-bottom: 10px;
-  }
-  
-  .subtext {
-    font-size: 14px;
-    margin: 5px 0;
-    text-align: center;
-  }
-
