--- calorieburningapp.client/src/Components/MainPage/index.js
@@ -1,1 +1,0 @@
-export { default } from './RegisterLogin';
