--- calorieburningapp.client/src/Pages/RegisterPage.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const RegisterPage: React.FC = () => {
-  return (
-    <div>
-      <h1>Register Page is here</h1>
-    </div>
-  );
-}
-
-export default RegisterPage;
